2d convex hull parsing (not yet plotting) + inputs

+++ inputs.py
@@ -106,7 +106,7 @@ done
         f.write(script)
 
 
-def phonopy(zpe_path):
+def phonopy(zpe_path, pressure):
     script = r"""#!/bin/sh
 
 for dir_name in `ls -d */`
@@ -133,6 +133,7 @@ ISMEAR = 1; SIGMA = 0.2
 LCHARG = .FALSE.
   NPAR = 4
 KSPACING = 0.314
+PSTRESS = 
 !
             cp ../CONTCAR POSCAR
             phonopy -d --dim 2 2 2 --tolerance=2e-1
@@ -167,6 +168,73 @@ done
 
 def gather(zpe_path):
     script = r"""#!/bin/sh
+
+for dir_name in `ls -d */`
+do
+    cd $dir_name
+    stat=$dir_name": "
+    if [ -d phonopy/ ]; then
+        cd phonopy/
+        echo $PWD
+        phonstat=""
+        if [ ! -f total_dos.pdf ]; then
+            for disp in `ls -d disp-*/`
+            do
+                cd $disp
+                if [ -f log ]; then
+                    err=`cat err`
+                    err=$err`grep 'BAD TERMINATION' log`
+                    err=$err`grep 'I REFUSE TO CONTINUE WITH THIS SICK JOB' log`
+                    log=`tail -n 1 log | awk '{print $2}'`
+                    if [ "$log" == "F=" ]; then
+                        problems=`grep 'very serious problems' log`
+                        if [ "$problems" == "" ]; then
+                            phonstat=$phonstat""
+                        else
+                            phonstat=$phonstat"problems in $disp "
+                            #sed -i "s/NPAR = 4/NPAR = 8/g" INCAR
+                            #sed -i "s/NPAR = 8/NPAR = 16/g" INCAR
+                            #sed -i "s/NPAR = 16/NPAR = 8/g" INCAR
+                            #sed -i "s/SBATCH -n 8/SBATCH -n 16/g" script.sh
+                            #sed -i "s/SBATCH -n 16/SBATCH -n 32/g" script.sh
+                            #sbatch script.sh
+                        fi
+                    else
+                        if [ "$err" != "" ]; then
+                            phonstat=$phonstat" error in $disp"
+                            #sed -i "s/NPAR = 4/NPAR = 8/g" INCAR
+                            #sed -i "s/NPAR = 8/NPAR = 16/g" INCAR
+                            #sed -i "s/NPAR = 16/NPAR = 8/g" INCAR
+                            #sed -i "s/SBATCH -n 8/SBATCH -n 16/g" script.sh
+                            #sed -i "s/SBATCH -n 16/SBATCH -n 32/g" script.sh
+                            #sbatch script.sh
+                        else
+                            phonstat=$phonstat"$disp still working "
+                        fi
+                    fi
+                else
+                    phonstat=$phonstat"$disp not yet started "
+                fi
+                cd ..
+            done
+        fi
+        cd ..
+    else
+        stat=$stat"no phonopy dir created yet"
+    fi
+    if [ "$phonstat" == "" ]; then
+        phonstat="all done"
+        cd phonopy
+        if [ ! -f FORCE_SETS ]; then
+            echo "creating FORCE_SETS"
+            disp_s=`ls -d disp-*/ | head -n 1`
+            disp_f=`ls -d disp-*/ | tail -n 1`
+            disp_s="${disp_s/disp-/}"
+            disp_f="${disp_f/disp-/}"
+            disp_s="${disp_s////}"
+            disp_f="${disp_f////}"
+            disp_range=disp-\{${disp_s}..${disp_f}\}
+            cat>phon.sh<<!
 phonopy -f $disp_range/vasprun.xml
 !
             chmod 777 phon.sh
@@ -191,53 +259,15 @@ MP = 31 31 31
                 echo "creating thermal_properties.yaml"
                 phonopy -t -p mesh.conf --tolerance=$tol --tmax 2000 > thermal.log
         fi
-        if [ ! -f pdos.conf ]; then
-                echo "creating pdos.conf"
-                atom_name=`sed '6q;d' POSCAR`
-                cat>pdos.conf <<!
-ATOM_NAME = $atom_name
-DIM = $dim
-PDOS = AUTO
-!
-        fi
-        if [ ! -f partial_dos.pdf ]; then
-                echo "creating partial_dos.pdf"
-                phonopy -s -p pdos.conf --tolerance=$tol
-        fi
-	if [ ! -f band.conf ]; then
-                echo "creating band.conf"
-                atom_name=`sed '6q;d' POSCAR`
-                cat>band.conf <<!
-ATOM_NAME = $atom_name
-DIM = $dim
-BAND = AUTO
-!
-        fi
-        if [ ! -f phonon_band.dat ]; then
-                 echo "creating phonon_band.dat"
-                 phonopy -p band.conf --tolerance=$tol
-        fi
-        if [ ! -f band.pdf ]; then
-                 phonopy -p -s band.conf --tolerance=$tol
-        fi
-        if [ ! -f band-pdos.conf ]; then
-                echo "creating band-pdos.conf"
-                atom_name=`sed '6q;d' POSCAR`
-                cat>band-pdos.conf <<!
-ATOM_NAME = $atom_name
-DIM = $dim
-BAND = AUTO
-PDOS = AUTO
-!
-        fi
-        if [ ! -f band_dos.pdf ]; then
-                 phonopy -p -s band-pdos.conf --tolerance=$tol
-        fi
         if [ ! -f phonon_band.pdf ]; then
                 echo "creaing phonon_band.pdf"
-                sumo-phonon-bandplot --dim $dim --height 6 --width 9 --dos total_dos.dat --symprec=$tol
+                python ~/cms-scripts/phonon-plots/config.py
+                phonopy -p band.conf
+                python ~/cms-scripts/phonon-plots/plot.py
         fi
-        rm mesh.yaml
+        if [ -f mesh.yaml ]; then
+                rm mesh.yaml
+        fi
         cd ..
     fi
     printf "$stat $phonstat \n"
